Enforced RBAC Admin-only Delete functionality in ticketController.js; Configured Dashboard.jsx to load the existing tickets upon loading the application for the first time without having to refresh the page; Set the default role in Register.jsx to STANDARD instead of non-existent role USER; fixed value assignments of role options in the jsx

+++ soc-backend/src/controllers/ticketController.js
@@ -127,6 +127,10 @@ const updateTicket = async (req, res)=>{
 //4. DELETE A TICKET
 const deleteTicket = async (req, res)=>{
     try{
+        // 1. Enforce RBAC on the backend!
+        if (req.user.role !== 'ADMIN') {
+            return res.status(403).json({ error: "Unauthorized. Only admins can delete tickets." });
+        }
         const {id} = req.params;
         await prisma.ticket.delete({
             where:{id: parseInt(id)}
